requirements.txt
fastapi
uvicorn
openai
python-dotenv
pypdf
python-docx
langchain
langchain-community
faiss-cpu
sentence-transformers
slowapi
pymongo
playwright
Pillow
httpx